Match credit-array indices against impressions in descending order
This makes the credit array resemble an infinite list that is truncated
at the point at which its values become 0.

For example, [.5, 0.25, 0.25] now means that the last 3 impressions will
be matched, and the last one will be given half of the credit, while the
next two will receive one quarter each.